feat: add perception radius configuration and refactor genome management
- Introduced a new `perception_radius` setting in `config.yaml` and `SimulationConfig`, enhancing agent environmental awareness.
- Refactored the `GenomeId` class, moving it to `database.data_types.py` for better organization and clarity in genetic tracking.
- Updated `BaseAgent` to utilize the new perception radius in its perception calculations, improving the agent's interaction with its environment.
- Cleaned up the `base_agent.py` file by removing deprecated code and ensuring consistent imports, enhancing overall code maintainability.

These changes significantly improve the configuration and management of agent perception and genetics, facilitating more sophisticated interactions within the simulation environment.
