+++ core/config.py
@@ -56,7 +56,7 @@ class SimulationConfig:
     offspring_cost: int = 6
     min_reproduction_resources: int = 10
     offspring_initial_resources: int = 5
-
+    perception_radius: int = 2
     # Agent type ratios
     agent_type_ratios: Dict[str, float] = field(
         default_factory=lambda: {
